// sanity.types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;

// Source: schema.json
export type FranchisePackage = {
  _id: string;
  _type: "franchisePackage";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  price?: string;
  priceNote?: string;
  features?: Array<string>;
  featured?: boolean;
  order?: number;
};

export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type BlockContent = Array<
  | {
      children?: Array<{
        marks?: Array<string>;
        text?: string;
        _type: "span";
        _key: string;
      }>;
      style?: "normal" | "h1" | "h2" | "h3" | "h4" | "blockquote";
      listItem?: "bullet";
      markDefs?: Array<{
        href?: string;
        _type: "link";
        _key: string;
      }>;
      level?: number;
      _type: "block";
      _key: string;
    }
  | {
      asset?: SanityImageAssetReference;
      media?: unknown;
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      alt?: string;
      _type: "image";
      _key: string;
    }
>;
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export type POST_BY_SLUG_QUERY_RESULT = {
  _id: string;
  title: string | null;
  slug: Slug | null;
  mainImage: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    _type: "image";
  } | null;
  publishedAt: string | null;
  body: BlockContent | null;
  author: {
    name: string | null;
    image: {
      asset?: SanityImageAssetReference;
      media?: unknown;
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      _type: "image";
    } | null;
  } | null;
  categories: Array<{
    title: string | null;
  }> | null;
} | null;

// Source: sanity/lib/queries.ts
// Variable: ALL_POST_SLUGS_QUERY
// Query: *[_type == "post" && defined(slug.current)] {    "slug": slug.current  }
export type ALL_POST_SLUGS_QUERY_RESULT = Array<{
  slug: string | null;
}>;